fix(snapdoc): fix DocumentType serializer
It was subtly broken previously, publicId and systemId are set in Chrome even if the value is the
empty string. Thus, "IsString()" comparison is not enough. I also felt that it would look weird
to use a systemId without a publicId, but I'm not stuck on this, and this could change with
some spec research.