test(snapdoc): add tests for DocumentType serialization
This CL also adds some helpers for testing fixtures in browser code. I opted against using XHR to
grab the DOM nodes, because it's nice to just let the browser do it all on its own, so I'm using
iframes for now. This could change in the future, I'm not totally stuck on this.