refactor(snapdoc): make displacement text algorithm shared
I don't want to re-implement this for every single node type, that would suck.
It's nice to be able to use that where possible, by letting the DOM implementation
do the work for me. But, this might be removed later on, and it can't be applied
to all node types.