feat(snapdoc): support serializing text nodes
Test nodes are important. But this is actually probably broken, especially for old browsers. I won't
know for sure unless I can run tests for this elsewhere.

To be honest, creating a snapshot of the page from what the browser sees is a bit silly, because we
can't deal with the raw content. But whatever, it's kinda fun.