feat(snapdoc): support serializing entire documents
This actually works better than I expected it to, I thought it might serialize the <head>
element which is absent from the actual test document. So at least in Chromes case, the
browser is a bit more clever.

But this is still probably not going to behave consistently on FireFox. I think the next
step (before worrying about things like document fragments), is to test it on another browser.